\name{catcheffortwb-package}
\alias{catcheffortwb-package}
\alias{catcheffortwb}
\docType{package}
\title{
\packageTitle{catcheffortwb}
}
\description{
\packageDescription{catcheffortwb}
}
\details{
  This package contains the data required to fit the catch-effort models
  described in Vajas et al. (in prep)
}
\author{
\packageAuthor{catcheffortwb}

Maintainer: \packageMaintainer{catcheffortwb}
}
\references{
  Vajas, P., Calenge, C. etc. XXX, in prep.
}
\keyword{ package }
\keyword{ model }
\seealso{
  \code{\link[<pkg>:<pkg>-package]{fitCatchEffort}}
}
