\name{catcheffortwb-package}
\alias{catcheffortwb-package}
\alias{catcheffortwb}
\docType{package}
\title{
\packageTitle{catcheffortwb}
}
\description{
\packageDescription{catcheffortwb}
}
\details{
  This package contains the data and the function required to fit the catch-effort models
  described in Vajas et al. (in prep).
}
\author{
\packageAuthor{catcheffortwb}

Maintainer: \packageMaintainer{catcheffortwb}
}
\references{
  Vajas, P., Calenge, C., Gamelon, M., Girard, F., Melac, O., Chandosne, C., Richard, E., Said, S., Baubet, Eric. in prep : Catch-effort model used as a management tool in exploited populations: wild boar as a case study.
}
\keyword{ package }
\keyword{ model }
\seealso{
  \code{\link[<pkg>:<pkg>-package]{fitCatchEffort}}
}
